#6539, #6534, #6541, #6544, volume capacity license
* #6534,update column source in cube build

* #6544,update history usage

* #6539,calculate lookup table row count

* #6541, capacity license rest apis

* minor eclipse file

* #6541 refactor, 1) no depend on hive-src; 2) SourceUsageParams=>SourceUsageRecord; 3) many FIXME #6541 markers

* minor,fix remove 0 capacity project

+++ core-dictionary/src/main/java/org/apache/kylin/dict/lookup/SnapshotTableV1.java
@@ -297,7 +297,7 @@ public class SnapshotTableV1 extends SnapshotTable {
     }
 
     @Override
-    public int getRowCount() {
+    public long getRowCount() {
         return rowIndices.size();
     }
